requirements.txt
Flask
Werkzeug
Pillow
numpy
nltk
python-docx
